CryptoPkg: Fix integer overflow
(cherry picked from commit 3c9e2f239a38590b4e3a8c1ec2304227f2af0103)

SECSPERDAY is 86400 which exceeds the limit of a UINT16 which is 65536.
Therefore DayRemainder cannot use UINT16. This patch makes it UINT32.

Cc: Ard Biesheuvel <ardb+tianocore@kernel.org>
Cc: Jordan Justen <jordan.l.justen@intel.com>
Cc: Laszlo Ersek <lersek@redhat.com>
Cc: Anthony Perard <anthony.perard@citrix.com>
Cc: Julien Grall <julien@xen.org>

(cherry picked from commit 3c9e2f239a38590b4e3a8c1ec2304227f2af0103)

+++ CryptoPkg/Library/BaseCryptLib/SysCall/TimerWrapper.c
@@ -118,7 +118,7 @@ gmtime (
 {
   struct tm  *GmTime;
   UINT16     DayNo;
-  UINT16     DayRemainder;
+  UINT32     DayRemainder;
   time_t     Year;
   time_t     YearNo;
   UINT16     TotalDays;
@@ -136,7 +136,7 @@ gmtime (
   ZeroMem ((VOID *)GmTime, (UINTN)sizeof (struct tm));
 
   DayNo        = (UINT16)(*timer / SECSPERDAY);
-  DayRemainder = (UINT16)(*timer % SECSPERDAY);
+  DayRemainder = (UINT32)(*timer % SECSPERDAY);
 
   GmTime->tm_sec  = (int)(DayRemainder % SECSPERMIN);
   GmTime->tm_min  = (int)((DayRemainder % SECSPERHOUR) / SECSPERMIN);
